Operations can't timeout when writing to the write buffer
We implemented a mechanism for operations to be able to timeout
while waiting to be written to a socket, but failed to take into
account the fact that an item can timeout while in the middle
of writing to that buffer.

This commit adds a new OperationState called WRITE_QUEUED that
specifies that the operation is waiting to be written to the
write buffer. When an operation begins writing to the write
buffer it goes into the WRITING state. When an operation is in
the WRITING state it cannot time out

Change-Id: I36eac0b5ca675ebe76ffe97e56e5fcf90b0916b2
Reviewed-on: http://review.couchbase.org/9223
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>